+++ nrv/utils/cell/CL_simulations.py
@@ -167,6 +167,148 @@ def axon_AP_threshold(axon: axon, amp_max: float, update_func: Callable,
                   + f" found in {Niter-1} iterations ({np.round(t_stop-t_start,2)}s).")
     return current_amp
 
+def axon_block_threshold(axon: axon, amp_max: float, update_func: Callable, AP_start: float,
+                      t_sim: float = 10, tol: float = 1, freq: float|None = None, verbose: bool = True, 
+                      args_update = dict|None, save_path: str|None = None, **kwargs) -> np.float64 :
+    """
+    Find the block threshold of an axon with arbitrary stimulation settings.
+
+    Parameters
+    ----------
+    axon : axon
+        axon object to simulation
+    amp_max : float
+        maximum amplitude for the binary search, in µA
+    update_func : Callable
+        Callable object to update the axon stimulation parameters between each binary search iteration
+    AP_start : float
+            timestamp of the test pulse start, in ms.
+    t_sim : float, optional
+        Simulation duration, in ms, by default 5
+    tol : float, optional
+        Search tolerance, in %, by default 1
+    freq : float, optional
+            Frequency of the stimulation, for KES block, by default None
+    verbose : bool, optional
+        Verbosity of the search, by default True
+    args_update : arg, optional
+        update_func arguments, by default None
+    save_path : str, optional
+        if not none, raster_plot and plot_x_t of axon results are saved in save_path
+
+    Returns
+    -------
+    current_amp : np.float64
+        The block threshold found with the binary search.
+    """
+
+    if "amp_tol_abs" in kwargs:
+        amp_tol_abs = kwargs.get("amp_tol_abs")
+    else:
+        amp_tol_abs = 0
+
+    if "amp_min" in kwargs:
+        amp_min = kwargs.get("amp_min")
+    else:
+        amp_min = 0
+    
+    amplitude_max_th = amp_max
+    amplitude_min_th = amp_min
+    amplitude_tol = tol
+
+    vm_key = "V_mem"
+    if (freq is not None):
+        vm_key += "_filtered"
+
+    # Dichotomy initialization
+    previous_amp = amp_min
+    delta_amp = np.abs(amp_max - amp_min)
+    current_amp = amp_max
+    Niter = 1
+    keep_going = 1
+    t_start = perf_counter()
+    
+    while keep_going:
+        if verbose and Niter == 1:
+            pass_info(f"Iteration {Niter}, Amp is {np.round(current_amp,2)}µA ...")
+            
+        update_func(axon,current_amp,**args_update)
+        results = axon.simulate(t_sim=t_sim)
+        is_blocked = results.is_blocked(AP_start=AP_start, freq=freq)
+        if save_path:
+            fig, ax = plt.subplots(1)
+            results.raster_plot(ax,key = vm_key)
+            ax.set_title(f"Block Amplitude: {np.round(current_amp,2)}µA")
+            fig.tight_layout()
+            fig_name = save_path + f"raster_plot_axon_{axon.ID}_iter_{Niter}.png"
+            fig.savefig(fig_name)
+            plt.close(fig)
+
+            fig, ax = plt.subplots(1)
+            results.plot_x_t(ax,key = vm_key)
+            ax.set_title(f"Block Amplitude: {np.round(current_amp,2)}µA")
+            fig.tight_layout()
+            fig_name = save_path + f"vmen_plot_axon_{axon.ID}_iter_{Niter}.png"
+            fig.savefig(fig_name)
+            plt.close(fig)
+
+        if verbose:
+            clear_prompt_line(1)
+
+        # post-process results
+        delta_amp = np.abs(current_amp - previous_amp)
+        #put in a function
+        if amp_tol_abs > 0:
+            if delta_amp < amp_tol_abs:
+                keep_going = 0
+            else:
+                keep_going = 1
+        else:
+            if current_amp > previous_amp:
+                current_tol = 100 * delta_amp / current_amp
+            else:
+                current_tol = 100 * delta_amp / previous_amp
+            if current_tol <= tol:
+                keep_going = 0
+            else:
+                keep_going = 1
+        previous_amp = current_amp
+        # test simulation results, update dichotomy
+        if is_blocked is None:
+            rise_warning("Failed to evaluate block state of the axon! Consider changing the simulation parameters.")
+            return 0.0
+        
+        if is_blocked:
+            if current_amp == amp_min:
+                rise_warning("Minimum Stimulation Current is too High!")
+                return 0.0
+            if verbose:
+                pass_info(f"Iteration {Niter}, Amp is {np.round(current_amp,2)}µA"+ 
+                          f" ({np.round(current_tol,2)}%)... AP Blocked! (in {np.round(results["sim_time"],3)}s)")
+                #pass_info("... Spike triggered")
+            amplitude_max_th = previous_amp
+            current_amp = (delta_amp / 2) + amplitude_min_th
+        else:
+            if current_amp == amp_max:
+                rise_warning("Maximum Stimulation Current is too Low!")
+                return 0.0
+            if verbose:
+                pass_info(f"Iteration {Niter}, Amp is {np.round(current_amp,2)}µA"+
+                          f" ({np.round(current_tol,2)}%)... AP Not Blocked! (in {np.round(results["sim_time"],3)}s)")
+            current_amp = amplitude_max_th - delta_amp / 2
+            amplitude_min_th = previous_amp
+
+        if previous_amp == amp_max:
+            current_amp = amp_min
+
+        Niter += 1
+    t_stop = perf_counter()
+    if verbose:
+        clear_prompt_line(1)
+        pass_info(f"Activation threshold is {np.round(current_amp,2)}µA ({np.round(current_tol,2)}%),"
+                  + f" found in {Niter-1} iterations ({np.round(t_stop-t_start,2)}s).")
+    return current_amp
+
 def firing_threshold_point_source(
     diameter,
     L,
@@ -847,6 +989,12 @@ def blocking_threshold_point_source(
             estimated threshold in uA
     """
 
+    rise_warning(
+        "DeprecationWarning: ",
+        "blocking_threshold_point_source is obsolete use axon_block_threshold instead"
+    )
+
+
     if "material" in kwargs:
         material = kwargs.get("material")
     else:
@@ -1125,6 +1273,11 @@ def blocking_threshold_from_axon(
     threshold       : float
             estimated threshold in uA
     """
+
+    rise_warning(
+        "DeprecationWarning: ",
+        "blocking_threshold_point_source is obsolete use blocking_threshold_from_axon instead"
+    )
 
     if "amp_min" in kwargs:
         amp_min = kwargs.get("amp_min")